Fix for broken image
I was reading your old posts for inspiration and noticed that this image was missing (I wanted to see it 😛) It looks like you tried to custom size it, but the markdown implementation being used doesn't support it(?)

Caveat - untested, because I was too lazy to build it locally, so I am not sure what it looks like, but should work I think. 

Also hi - long time no see 🙂 hope you're doing well!